Encryption

+++ cache_interceptor/src/main/java/uk/co/glass_software/android/cache_interceptor/response/CachedResponse.java
@@ -14,7 +14,7 @@ public abstract class CachedResponse<E extends Exception & Function<E, Boolean>,
     private ResponseMetadata<R, E> metadata;
     
     @Override
-    public int getTtlInMinutes() {
+    public float getTtlInMinutes() {
         return DEFAULT_TTL_IN_MINUTES;
     }
